add binary instructions and update tests for new prints

+++ compiler/mir/segment.h
@@ -101,8 +101,8 @@ class Segment final {
     std::unordered_map<int16_t, ConstantInt *> m_I16Pool = {};
     std::unordered_map<int32_t, ConstantInt *> m_I32Pool = {};
     std::unordered_map<int64_t, ConstantInt *> m_I64Pool = {};
-    std::unordered_map<uint32_t, ConstantFP *> m_F32Pool = {};
-    std::unordered_map<uint64_t, ConstantFP *> m_F64Pool = {};
+    std::unordered_map<float, ConstantFP *> m_F32Pool = {};
+    std::unordered_map<double, ConstantFP *> m_F64Pool = {};
 
 public:
     Segment(const Target &T);
